doc: add more context about security in the pipeline and operator
The current documentation doesn't cover all the security scan and measures
that has been taken in the last year, so we added that in the documentation
and also, a little section explaining how are the operator permissions
managed in the different way the operator supports and some
recommendations about how to deploy it. In the new section we also
explain more about he cluster wide permissions that are mandatory for the
operator to work.

Closes #4348

Signed-off-by: Jonathan Gonzalez V <jonathan.gonzalez@enterprisedb.com>